requirements.txt
asgiref==3.2.10
astroid==2.3.2
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.3
django-allauth==0.38.0
django-cors-headers==3.5.0
django-filter==2.1.0
django-multiselectfield==0.1.12
django-rest-auth==0.9.3
djangorestframework==3.9.1
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
Markdown==3.0.1
mccabe==0.6.1
oauthlib==3.1.0
psycopg2==2.7.3.1
pylint==2.4.3
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.6
wrapt==1.11.2
